Extract parseExtensionSection to misc
